fitness from float to double

+++ src/evolution/Community.java
@@ -3,7 +3,7 @@ package evolution;
 import game.Tickable;
 
 public interface Community<T> extends Alive, Tickable, Combinable<T>, Mutable {
-    float fitness();
+    double fitness();
     int getScore();
     T copy();
     T newIndividual();
